move the NEON counting kernel out of src/generic into src/arm64
src/generic/utf8_validation/utf8_validator.h carried a NEON implementation
(int8x16_t, vaddlvq_s8, vcgtq_u8, vdupq_n_s8) behind #if
SIMDUTF_IMPLEMENTATION_ARM64, in a header shared by every kernel. Move
neon_counter and the validator to src/arm64/arm_validate_utf8_with_counts.cpp,
renamed arm_validate_utf8_with_counts to match its arm_* neighbours, and have
arm64's implementation.cpp call it directly. The moved code is unchanged.

The generic const char* wrapper becomes unused on arm64 once that call site
goes away, which trips -Wunused-function under the -Werror macOS CI job, so it
is now simdutf_really_inline.

utf8_counters kept an arm64 branch selecting sum_bytes() over to_bitmask() +
count_ones. That branch is unreachable now that arm64 has its own counting
loop, so it is removed; both generic files are free of architecture checks.

Pure code motion: the disassembly of
simdutf::arm64::implementation::validate_utf8_with_counts is byte-identical
before and after, as is the whole simdutf.cpp.o.
