+++ src/arm64/implementation.cpp
@@ -170,6 +170,11 @@ convert_utf8_1_to_2_byte_to_utf16(uint8x16_t in, size_t shufutf8_idx) {
   #include "generic/utf8_validation/utf8_lookup4_algorithm.h"
   #include "generic/utf8_validation/utf8_validator.h"
 #endif // SIMDUTF_FEATURE_UTF8 || SIMDUTF_FEATURE_DETECT_ENCODING
+#if SIMDUTF_FEATURE_UTF8
+  // needs utf8_checker and buf_block_reader, hence included after the generic
+  // validator
+  #include "arm64/arm_validate_utf8_with_counts.cpp"
+#endif // SIMDUTF_FEATURE_UTF8
 
 #if SIMDUTF_FEATURE_ASCII
   #include "generic/ascii_validation.h"
@@ -1117,7 +1122,7 @@ implementation::count_utf8(const char *input, size_t length) const noexcept {
 }
 simdutf_warn_unused utf8_result implementation::validate_utf8_with_counts(
     const char *buf, size_t len) const noexcept {
-  return arm64::utf8_validation::generic_validate_utf8_with_counts(buf, len);
+  return arm64::utf8_validation::arm_validate_utf8_with_counts(buf, len);
 }
 
 #endif // SIMDUTF_FEATURE_UTF8
